4.21 working i think

+++ controllers/blogs.js
@@ -38,9 +38,19 @@ blogsRouter.get('/', (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) =>  {
   const id = request.params.id
-
-  await Blog.findByIdAndRemove(id)
-  response.status(204).json()
+  const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const blog = await Blog.findById(id)
+  const user = await User.findById(decodedToken.id)
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(id)
+    response.status(204).json()
+  } else {
+    return response.status(401).json({ error: 'token does not match blog creators id' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
